feat: Add table information and status cards with detailed views
- Implemented TableInfoCard component to display table capacity, zone, and last update time.
- Created TableStatusCard component to show table status, including closing, reserved, and locked states.
- Developed VoucherSection component for managing order vouchers, including applying and removing vouchers.
- Introduced TableCard component for displaying individual table details and actions based on status.
- Added TableList component to manage and display a list of tables grouped by zones.
- Implemented TableLockDialog for locking tables with maintenance notes.
- Created TableZoneGroup to group tables by zone and display them in a structured layout.
