+++ src/apis/axiosConfig.ts
@@ -2,7 +2,6 @@ import axios, { type AxiosInstance, type AxiosError, type AxiosResponse, type In
 
 const apiUrl = import.meta.env.VITE_API_URL
 
-
 // Create Axios instance
 const api: AxiosInstance = axios.create({
   baseURL: apiUrl,
