+++ src/pages/InTables/components/tables-details/table-info-card.tsx
