+++ src/pages/InTables/components/tables-details/table-status-card.tsx
